refactor: useListNavigation custom hook을 생성하여 코드분리
